Add ability to delete locally stored files off our server.
This will happen when we delete the products from the admin page or we
edit a products image. It deletes the old image and stores the new one.
Also had to adjust the logic due to us slicing the image path when we
store it, so we need to use path.join to rejoin the original path that
is nested inside the src folder